refactoring

+++ GaelO-Flow/src/users/users.controller.ts
@@ -9,6 +9,7 @@ import {
   HttpException,
   UseInterceptors,
   UseGuards,
+  NotFoundException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { User } from './user.entity';
@@ -72,10 +73,13 @@ export class UsersController {
 
   @UseGuards(AdminGuard)
   @Delete('/:id')
-  @UseInterceptors(NotFoundInterceptor)
   async delete(@Param('id') id: number): Promise<void> {
-    const user = await this.UserService.findOne(id);
-    return await this.UserService.remove(id);
+    const existingUser = await this.UserService.isExistingUser(id);
+    if (existingUser) {
+      return await this.UserService.remove(id);
+    } else {
+      throw new NotFoundException('All the keys are required');
+    }
   }
 
   @UseGuards(AdminGuard)
